backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.3.0
djoser==2.2.0
drf-yasg==1.21.7
django-environ==0.11.2
Pillow==10.0.0
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.5.0